+++ src-ui/app/app.module.ts
@@ -34,7 +34,7 @@ import { SleepingPoseViewerComponent } from './components/sleeping-pose-viewer/s
 import { OscService } from './services/osc.service';
 import { OscAutomationsViewComponent } from './views/dashboard-view/views/osc-automations-view/osc-automations-view.component';
 import { SelectBoxComponent } from './components/select-box/select-box.component';
-import { TStringTranslatePipePipe } from './pipes/tstring-translate.pipe';
+import { TStringTranslatePipe } from './pipes/tstring-translate.pipe';
 import { OscScriptButtonComponent } from './components/osc-script-button/osc-script-button.component';
 import { OscScriptModalComponent } from './components/osc-script-modal/osc-script-modal.component';
 import { OscScriptCodeEditorComponent } from './components/osc-script-code-editor/osc-script-code-editor.component';
@@ -80,11 +80,10 @@ import { SleepModeForSleepDetectorAutomationService } from './services/sleep-det
 import { SleepDetectorCalibrationModalComponent } from './views/dashboard-view/views/sleep-detection-view/sleep-detector-calibration-modal/sleep-detector-calibration-modal.component';
 import { SleepDetectorEnableSleepModeModalComponent } from './views/dashboard-view/views/sleep-detection-view/sleep-detector-enable-sleepmode-modal/sleep-detector-enable-sleep-mode-modal.component';
 import { SettingsNotificationsTabComponent } from './views/dashboard-view/views/settings-view/settings-notifications-tab/settings-notifications-tab.component';
-import { DisplayBrightnessControlService } from './services/brightness-control/display-brightness/display-brightness-control.service';
+import { DisplayBrightnessControlService } from './services/brightness-control/display-brightness-control.service';
 import { BrightnessAutomationsViewComponent } from './views/dashboard-view/views/brightness-automations-view/brightness-automations-view.component';
 import { SliderSettingComponent } from './components/slider-setting/slider-setting.component';
 import { SliderComponent } from './components/slider/slider.component';
-import { DisplayBrightnessControlAutomationService } from './services/brightness-control/display-brightness/display-brightness-control-automation.service';
 import { EventLogService } from './services/event-log.service';
 import { debug } from 'tauri-plugin-log-api';
 import { EventLogComponent } from './components/event-log/event-log.component';
@@ -123,12 +122,10 @@ import { TurnOffLighthousesOnSteamVRStopAutomationService } from './services/pow
 import { ShutdownAutomationsViewComponent } from './views/dashboard-view/views/shutdown-automations-view/shutdown-automations-view.component';
 import { ShutdownAutomationsService } from './services/shutdown-automations.service';
 import { ShutdownSequenceOverlayComponent } from './components/shutdown-sequence-overlay/shutdown-sequence-overlay.component';
-import { DisplayBrightnessAutomationsTabComponent } from './views/dashboard-view/views/brightness-automations-view/tabs/display-brightness-automations-tab/display-brightness-automations-tab.component';
-import { ImageBrightnessAutomationsTabComponent } from './views/dashboard-view/views/brightness-automations-view/tabs/image-brightness-automations-tab/image-brightness-automations-tab.component';
-import { ImageBrightnessControlService } from './services/brightness-control/image-brightness/image-brightness-control.service';
-import { ImageBrightnessControlAutomationService } from './services/brightness-control/image-brightness/image-brightness-control-automation.service';
+import { ImageBrightnessControlService } from './services/brightness-control/image-brightness-control.service';
+import { BrightnessControlAutomationService } from './services/brightness-control/brightness-control-automation.service';
 import { DeveloperDebugModalComponent } from './components/developer-debug-modal/developer-debug-modal.component';
-import { DeveloperDebugService } from './services/developer-debug.service';
+import { DeveloperDebugService } from './services/developer-debug/developer-debug.service';
 import { MomentModule } from 'ngx-moment';
 import { IPCStateSyncService } from './services/ipc-state-sync.service';
 import { IPCService } from './services/ipc.service';
@@ -140,6 +137,11 @@ import { NotificationService } from './services/notification.service';
 import { WindowsPowerPolicyTabComponent } from './views/dashboard-view/views/power-automations-view/tabs/windows-power-policy-tab/windows-power-policy-tab.component';
 import { SetWindowsPowerPolicyOnSleepModeAutomationService } from './services/power-automations/set-windows-power-policy-on-sleep-mode-automation.service';
 import { SteamService } from './services/steam.service';
+import { BrightnessAutomationsTabComponent } from './views/dashboard-view/views/brightness-automations-view/tabs/brightness-automations-tab/brightness-automations-tab.component';
+import { TooltipDirective } from './directives/tooltip.directive';
+import { SimpleBrightnessControlService } from './services/brightness-control/simple-brightness-control.service';
+import { DebugSleepDetectionDebuggerComponent } from './components/developer-debug-modal/debug-sleep-detection-debugger/debug-sleep-detection-debugger.component';
+import { DebugBrightnessTestingComponent } from './components/developer-debug-modal/debug-brightness-testing/debug-brightness-testing.component';
 
 [localeEN, localeFR, localeCN_TW, localeNL, localeKO, localeJP, localeES, localeID].forEach(
   (locale) => registerLocaleData(locale)
@@ -160,6 +162,7 @@ export function createTranslateLoader(http: HttpClient) {
     DeviceListComponent,
     DeviceListItemComponent,
     VarDirective,
+    TooltipDirective,
     ImageFallbackDirective,
     AboutViewComponent,
     OverviewViewComponent,
@@ -176,7 +179,7 @@ export function createTranslateLoader(http: HttpClient) {
     SleepingPoseViewerComponent,
     OscAutomationsViewComponent,
     SelectBoxComponent,
-    TStringTranslatePipePipe,
+    TStringTranslatePipe,
     LocalizedDatePipe,
     ImageCachePipe,
     OscScriptButtonComponent,
@@ -202,8 +205,7 @@ export function createTranslateLoader(http: HttpClient) {
     GpuPowerlimitingPaneComponent,
     MsiAfterburnerPaneComponent,
     BrightnessAutomationsViewComponent,
-    DisplayBrightnessAutomationsTabComponent,
-    ImageBrightnessAutomationsTabComponent,
+    BrightnessAutomationsTabComponent,
     SliderSettingComponent,
     SliderComponent,
     EventLogComponent,
@@ -220,6 +222,8 @@ export function createTranslateLoader(http: HttpClient) {
     DeveloperDebugModalComponent,
     DotnetUpgradeModalComponent,
     WindowsPowerPolicyTabComponent,
+    DebugSleepDetectionDebuggerComponent,
+    DebugBrightnessTestingComponent,
   ],
   imports: [
     CommonModule,
@@ -238,7 +242,7 @@ export function createTranslateLoader(http: HttpClient) {
     }),
     NgPipesModule,
   ],
-  providers: [ThemeService],
+  providers: [ThemeService, TStringTranslatePipe],
 })
 export class AppModule {
   constructor(
@@ -258,6 +262,7 @@ export class AppModule {
     private imageCacheService: ImageCacheService,
     private displayBrightnessControlService: DisplayBrightnessControlService,
     private imageBrightnessControlService: ImageBrightnessControlService,
+    private simpleBrightnessControlService: SimpleBrightnessControlService,
     private systemTrayService: SystemTrayService,
     private eventLog: EventLogService,
     private lighthouseService: LighthouseService,
@@ -295,8 +300,7 @@ export class AppModule {
     // Shutdown automations
     private shutdownAutomationsService: ShutdownAutomationsService,
     // Brightness control automations
-    private displayBrightnessControlAutomationService: DisplayBrightnessControlAutomationService,
-    private imageBrightnessControlAutomationService: ImageBrightnessControlAutomationService,
+    private brightnessControlAutomationService: BrightnessControlAutomationService,
     // Render resolution automations
     private renderResolutionAutomationService: RenderResolutionAutomationService,
     // Chaperone fade dinstance automations
@@ -345,6 +349,7 @@ export class AppModule {
           this.displayBrightnessControlService.init(),
           this.imageBrightnessControlService.init(),
         ]);
+        await this.simpleBrightnessControlService.init();
         // Initialize IPC
         await this.ipcService.init();
         await this.ipcAppStateSyncService.init();
@@ -376,8 +381,7 @@ export class AppModule {
           // Invite automations
           this.inviteAutomationsService.init(),
           // Brightness automations
-          this.displayBrightnessControlAutomationService.init(),
-          this.imageBrightnessControlAutomationService.init(),
+          this.brightnessControlAutomationService.init(),
           // Resolution automations
           this.renderResolutionAutomationService.init(),
           // Fade distance automations